Cinemachine test + Misc Lemon tweaks
Added a (somewhat jenky) cinemachine camera to lemon, also adjusted the platforms in our jump test so he can make it through. Tweaked his character controller to reduce jitter when jumping next to an object, and added a "lockpoint" to the climb script. This snaps lemon to a position where it should be harder for him to climb at weird angles